Add work.html, form to create a work in database

+++ userboard/models.py
@@ -3,14 +3,15 @@ from django.contrib.auth.models import User
 from django import forms
 from django.contrib.auth.forms import UserCreationForm
 
+
 class Work(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
     name = models.CharField(max_length=32)
-    salary = models.PositiveSmallIntegerField()
-    tax_percent = models.PositiveSmallIntegerField()
-    extra_public_holiday_percent= models.PositiveSmallIntegerField()
-    extra_per_day = models.PositiveSmallIntegerField()
-    extra_additional_hour_percent = models.PositiveSmallIntegerField()
+    salary = models.DecimalField(decimal_places=2, max_digits=5)
+    tax_percent = models.DecimalField(decimal_places=2, max_digits=5)
+    extra_public_holiday_percent= models.DecimalField(decimal_places=2, max_digits=5)
+    extra_per_day = models.DecimalField(decimal_places=2, max_digits=5)
+    extra_additional_hour_percent = models.DecimalField(decimal_places=2, max_digits=5)
 
 class WorkHour(models.Model):
     work = models.ForeignKey(Work, on_delete=models.CASCADE)
